Created the custom hook and returning state to parent component
