+++ .eslintrc.json
@@ -5,17 +5,23 @@
     },
     "extends": [
         "eslint:recommended",
-        "plugin:react/recommended"
+        "plugin:react/recommended",
+        "plugin:react-hooks/recommended"
     ],
     "parserOptions": {
         "ecmaVersion": "latest",
         "sourceType": "module"
     },
     "plugins": [
-        "react"
+        "react",
+        "react-hooks"
     ],
     "rules": {
       "react/react-in-jsx-scope": "off",
-      "react/prop-types": [0, { "ignore": ["children"] }]
+      "react/prop-types": [0, { "ignore": ["children"] }],
+      "react-hooks/rules-of-hooks": "error",
+      "react-hooks/exhaustive-deps": ["warn", {
+        "additionalHooks": "(useMyCustomHook|useMyOtherCustomHook)"
+      }]
     }
 }
