+++ backend/data/user-places.json
@@ -1,1 +1,32 @@
-[]
+[
+  {
+    "id": "p3",
+    "title": "Himalayan Peaks",
+    "image": {
+      "src": "majestic-mountains.jpg",
+      "alt": "The sun setting behind snow-capped peaks of majestic mountains."
+    },
+    "lat": 27.9881,
+    "lon": 86.925
+  },
+  {
+    "id": "p14",
+    "title": "Kerala Backwaters",
+    "image": {
+      "src": "kerala-backwaters.jpg",
+      "alt": "Tranquil waters and lush greenery of the Kerala backwaters."
+    },
+    "lat": 9.4981,
+    "lon": 76.3388
+  },
+  {
+    "id": "p13",
+    "title": "Taj Mahal",
+    "image": {
+      "src": "taj-mahal.jpg",
+      "alt": "The iconic Taj Mahal reflecting in its surrounding waters during sunrise."
+    },
+    "lat": 27.1751,
+    "lon": 78.0421
+  }
+]
